+++ src/models/Settings.php
@@ -73,6 +73,24 @@ class Settings extends Model
     public $deleteTypes;
 
     /**
+     * @var bool
+     * @since 2.5.0
+     */
+    public bool $rulesAreWhitelist = false;
+
+    /**
+     * @var array
+     * @since 2.5.0
+     */
+    public $ignoreUserGroups = '';
+
+    /**
+     * @var bool
+     * @since 2.5.0
+     */
+    public bool $ignoreAdmins = false;
+
+    /**
      * @var array
      */
     public $ignoreRules = [
@@ -86,26 +104,23 @@ class Settings extends Model
     /**
      * Is a log type ignored by the set of rules
      *
-     * @param  string  $handle
+     * @param  string  $type
      * @return boolean
      */
-    public function isTypeIgnored(string $handle): bool
+    public function isTypeIgnored(string $type): bool
     {
+        if (!$this->userMatches()) {
+            return true;
+        }
         if (!$this->ignoreRules) {
-            return false;
+            return $this->rulesAreWhitelist;
         }
         foreach ($this->ignoreRules as $rule) {
-            if ($rule['active'] and ($rule['type'] == $handle or !$rule['type'])) {
-                if (!$rule['request'] or
-                    ($rule['request'] == 'yaml' and \Craft::$app->projectConfig->isApplyingExternalChanges) or
-                    ($rule['request'] == 'console' and \Craft::$app->request->isConsoleRequest) or
-                    ($rule['request'] == 'cp' and \Craft::$app->request->isCpRequest) or
-                    ($rule['request'] == 'site' and \Craft::$app->request->isSiteRequest)) {
-                    return true;
-                }
+            if ($this->ruleMatches($rule, $type)) {
+                return !$this->rulesAreWhitelist;
             }
         }
-        return false;
+        return $this->rulesAreWhitelist;
     }
 
     /**
@@ -135,5 +150,52 @@ class Settings extends Model
             return true;
         }
         return in_array($type, is_array($this->deleteTypes) ? $this->deleteTypes : []);
+    }
+
+    /**
+     * Does a rule match
+     * 
+     * @param array $rule
+     * @param string $type
+     * @return bool
+     * @since 2.5.0
+     */
+    protected function ruleMatches(array $rule, string $type): bool
+    {
+        if ($rule['active'] and ($rule['type'] == $type or !$rule['type'])) {
+            return (
+                !$rule['request'] or
+                ($rule['request'] == 'yaml' and \Craft::$app->projectConfig->isApplyingExternalChanges) or
+                ($rule['request'] == 'console' and \Craft::$app->request->isConsoleRequest) or
+                ($rule['request'] == 'cp' and \Craft::$app->request->isCpRequest) or
+                ($rule['request'] == 'site' and \Craft::$app->request->isSiteRequest)
+            );
+        }
+        return false;
+    }
+
+    /**
+     * Does the user match
+     * 
+     * @return bool
+     * @since 2.5.0
+     */
+    protected function userMatches(): bool
+    {
+        $user = \Craft::$app->getUser()->getIdentity();
+        if ($user) {
+            if ($this->ignoreAdmins and $user->admin) {
+                return false;
+            }
+            if (!$this->ignoreUserGroups) {
+                return true;
+            }
+            foreach ($this->ignoreUserGroups as $group) {
+                if ($user->isInGroup($group)) {
+                    return false;
+                }
+            }
+        }
+        return true;
     }
 }